ProminentProjectAction does not work as expected for MavenModule.  This is because getProminentActions() is a method of Project and not AbstractProject.
To my mind, getProminentActions should be a method of AbstractProject.  If specific project types (i.e. MavenModuleSet or MatrixProject) don't want the concept of ProminentProjectAction they can override the method to return an empty set, or but this default behaviour of filtering all actions and returning only the ones that implement ProminentProjectAction seems best placed in AbstractProject.

See http://www.nabble.com/Getting-ProminentProjectAction-to-work-to14724625.html