[HUDSON-2637] Copy of some of the changes from the supplied code... I'd prefer changes as diffs rather than the whole file as it can get tricky to apply multiples.
Also the rest of the attached changes makes invalid assumptions about how accurev can apply changes to a stream (i.e. it will miss some actual changes)

I agree that the areas touched need improvement, but the changes in this patch would make the detection less reliable IMHO.

It is better to trigger an unnecessary build than to fail to trigger a necessary one.