Added ability to construct a set of sensitive build variables.
Added AbstractBuild.getSensitiveBuildVariables which returns a set of variable
names which contain sensitive information that should not be exposed in
display output. This set is expected to consist of a subset of the keys in the
map returned by getBuildVariables.

Added method to allow ParamaterValues to indicate whether or not they have
sensitive information. Make PasswordParameterValue use this to indicate that
it does.

Give BuildWrappers a way to indicate what variables they contributed to a build
should be considered sensitive.